fix request regis
