--- app/utils/security.py
@@ -1,4 +1,0 @@
-import hashlib
-
-def hash_password(password: str) -> str:
-    return hashlib.sha256(password.encode()).hexdigest()
